requirements.txt
# Core dependencies
torch==2.3.0
torchvision
torchaudio
diffusers==0.27.0
transformers==4.36.0
accelerate==0.25.0
peft==0.7.1

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
svgwrite==1.4.3
svglib==1.5.1
shapely==2.0.2
reportlab==4.0.7

# OCR
pytesseract==0.3.10
pdf2image==1.16.3
easyocr==1.7.0

# Optimization
ortools==9.8.3296

# UI
streamlit==1.28.0

# Data processing
pandas==2.1.3
numpy==1.24.4

# PDF processing
PyPDF2==3.0.1
camelot-py[cv]==0.11.0

# Additional utilities
python-dotenv==1.0.0
